feat: Clarity 분리
- 별도의 파일로 관리
